ctrl to lf, fix bug on call_command
