+++ docker-compose.yml
@@ -30,6 +30,7 @@ services:
     depends_on:
       - database
       - redis
+        
 
   nginx:
     build: ./nginx
