+++ video_app/api_views/video_detail_api.py
@@ -13,8 +13,8 @@ from media_stream.utils import custom_exceptions as ce
 from media_stream.utils.custom_pagination import CustomPagination
 from media_stream.utils.standard_response import get_response_structure
 
-from ..models import VideoModel, VideoProcessingStatus
-from ..serializers import VideoSerializer
+# from ..models import VideoModel, VideoProcessingStatus
+# from ..serializers import VideoSerializer
 
 logger = logging.getLogger('video_app')
 
@@ -23,38 +23,38 @@ class VersioningConfig(NamespaceVersioning):
     allowed_versions = ['v1']
     version_param = 'version'
 
-class VideoDetailApi(APIView):
-    permission_classes = (AllowAny,)
-    versioning_class = VersioningConfig
-    page_size = 12
+# class VideoDetailApi(APIView):
+#     permission_classes = (AllowAny,)
+#     versioning_class = VersioningConfig
+#     page_size = 12
 
-    def get(self, request, **kwargs):
-        try :
-            if request.version == 'v1':
-                video = VideoModel.filter_cache(
-                    title_slug = kwargs['video_title'],
-                    processing_status= VideoProcessingStatus.FINISHED
-                ).first()
+#     def get(self, request, **kwargs):
+#         try :
+#             if request.version == 'v1':
+#                 video = VideoModel.filter_cache(
+#                     title_slug = kwargs['video_title'],
+#                     processing_status= VideoProcessingStatus.FINISHED
+#                 ).first()
                     
-                if video is None:
-                    raise ce.ResourceNotFound
+#                 if video is None:
+#                     raise ce.ResourceNotFound
 
-                video = VideoSerializer(
-                    instance= video, many= False
-                ).data
+#                 video = VideoSerializer(
+#                     instance= video, many= False
+#                 ).data
 
-                response = get_response_structure(
-                    status_code= status.HTTP_200_OK,
-                    success= True,
-                    message= message.DATA_FETHCED_SUCCESSFULLY,
-                    data= video
-                )             
-                return Response(response, status = response['status_code'])
+#                 response = get_response_structure(
+#                     status_code= status.HTTP_200_OK,
+#                     success= True,
+#                     message= message.DATA_FETHCED_SUCCESSFULLY,
+#                     data= video
+#                 )             
+#                 return Response(response, status = response['status_code'])
 
-        except ce.ResourceNotFound as _:
-            raise ce.ResourceNotFound
+#         except ce.ResourceNotFound as _:
+#             raise ce.ResourceNotFound
 
-        except :
-            logger.error(traceback.format_exc())
-            raise ce.InternalServerError
+#         except :
+#             logger.error(traceback.format_exc())
+#             raise ce.InternalServerError
