+++ video_app/api_views/video_list_api.py
@@ -11,8 +11,8 @@ from rest_framework.filters import SearchFilter
 from media_stream.utils import custom_exceptions as ce 
 from media_stream.utils.custom_pagination import CustomPagination
 
-# from ..models import TvChannelModel
-# from ..serializers import TvSerializer
+# from ..models import VideoModel, VideoProcessingStatus
+# from ..serializers import VideoSerializer
 
 logger = logging.getLogger('video_app')
 
@@ -21,20 +21,22 @@ class VersioningConfig(NamespaceVersioning):
     allowed_versions = ['v1']
     version_param = 'version'
 
-# class TvListApi(generics.ListAPIView):
+# class VideoListApi(generics.ListAPIView):
 #     permission_classes = (AllowAny,)
 #     versioning_class = VersioningConfig
-#     serializer_class = TvSerializer
+#     serializer_class = VideoSerializer
 #     pagination_class = CustomPagination
 #     filter_backends = (DjangoFilterBackend, SearchFilter, OrderingFilter)
 #     filterset_fields = (
-#         'channel_name_slug', 'language__name_slug', 
+#         'title_slug', 'language__name_slug', 'genre__name_slug', 
 #         'category__name_slug'
 #     )
-#     search_fields = ('channel_name', )
+#     search_fields = ('title', 'language__name_slug')
 #     ordering = ('-id', )
-#     queryset = TvChannelModel.cache_all()
-    
+#     queryset = VideoModel.filter_cache(
+#         processing_status= VideoProcessingStatus.FINISHED
+#     )
+
 #     def get(self, request, *args, **kwargs):
 #         try :
 #             if request.version == 'v1':
