+++ video_app/forms.py
@@ -17,22 +17,6 @@ class CustomUserChangeForm(UserChangeForm):
         model = models.CustomUser
         fields = ('email',)
 
-class LanguageForm(forms.ModelForm):
-    
-    class Meta: 
-        model = models.LanguageModel
-        fields = '__all__'
-
-class VideoForm(forms.ModelForm):
-    
-    # validators=[validate_even]
-
-    class Meta: 
-        model = models.VideoModel
-        fields = '__all__'
-        widgets = {
-            'video_file_path': forms.FileInput(attrs={'accept':'.mp4'})
-        }
 
 class ContactUsForm(forms.ModelForm):
     full_name = forms.CharField(label='Full name', max_length=50,
@@ -55,12 +39,31 @@ class ContactUsForm(forms.ModelForm):
         model = models.ContactUsModel
         fields = '__all__'
 
-class CategoryForm(forms.ModelForm):
+class LanguageForm(forms.ModelForm):
+    
     class Meta: 
-        model = models.CategoryModel
+        model = models.LanguageModel
         fields = '__all__'
+
 
 class GenreForm(forms.ModelForm):
     class Meta: 
         model = models.GenreModel
         fields = '__all__'
+
+
+class CategoryForm(forms.ModelForm):
+    class Meta: 
+        model = models.CategoryModel
+        fields = '__all__'
+
+
+class VideoForm(forms.ModelForm):
+    class Meta: 
+        model = models.VideoModel
+        fields = '__all__'
+        widgets = {
+            'video_file_path': forms.FileInput(attrs={'accept':'.mp4'})
+        }
+
+
