+++ video_app/views.py
@@ -7,8 +7,7 @@ from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 # from django.db.models import Q
 
 from video_app.forms import ContactUsForm
-from video_app.models import VideoProcessingStatus, VideoModel, TvChannelModel
-
+from video_app.models import TvChannelModel, VideoProcessingStatus, VideoModel
 
 va_logger = logging.getLogger('video_app')
 
@@ -41,6 +40,55 @@ class ContactUsView(View):
             context['form'] = contact_us
         return render(request, self.template_name, context= context)
 
+
+
+class TvListView(View):
+    template_name = 'tv/main.html'
+    page_size = 12
+
+    def get(self, request, *args, **kwargs):
+        context = {}
+        
+        search_channel= request.GET.get("search-channel")
+        if search_channel:
+            channels = TvChannelModel.filter_cache(
+                channel_name__icontains= search_channel
+            )
+        else:
+            channels = TvChannelModel.cache_all()
+        
+        paginator = Paginator(channels, TvListView.page_size)
+        page = request.GET.get('page', 1)
+        try:
+            channels = paginator.page(page)
+        except PageNotAnInteger:
+            channels = paginator.page(1)
+        except EmptyPage:
+            channels = paginator.page(paginator.num_pages)
+
+        if channels:
+            context['channels'] = channels
+        elif search_channel:
+            context['error_message'] = f'Sorry could not find {search_channel}'
+        return render(request, self.template_name, context= context)
+
+class TvDetailView(View):
+    template_name = 'tv/watch_channel.html'
+
+    def get(self, request, *args, **kwargs):
+        channel = None
+        try:
+            channel = TvChannelModel.filter_cache(
+                channel_name_slug = kwargs['channel_name_slug']
+                ).first()
+            
+        except :
+            va_logger.error(traceback.format_exc())
+            # TODO : renders ops page
+        
+        return render(request, self.template_name, context= {
+            'channel' : channel
+        })
 
 class VideoListView(View):
     template_name = 'videos/main.html'
@@ -131,51 +179,3 @@ class VideoCaterogry(View):
         elif search_title:
             context['message'] = f'Sorry could not find {search_title}'
         return render(request, self.template_name, context= context)
-
-class TvListView(View):
-    template_name = 'tv/main.html'
-    page_size = 12
-
-    def get(self, request, *args, **kwargs):
-        context = {}
-        
-        search_channel= request.GET.get("search-channel")
-        if search_channel:
-            channels = TvChannelModel.filter_cache(
-                channel_name__icontains= search_channel
-            )
-        else:
-            channels = TvChannelModel.cache_all()
-        
-        paginator = Paginator(channels, TvListView.page_size)
-        page = request.GET.get('page', 1)
-        try:
-            channels = paginator.page(page)
-        except PageNotAnInteger:
-            channels = paginator.page(1)
-        except EmptyPage:
-            channels = paginator.page(paginator.num_pages)
-
-        if channels:
-            context['channels'] = channels
-        elif search_channel:
-            context['error_message'] = f'Sorry could not find {search_channel}'
-        return render(request, self.template_name, context= context)
-
-class TvDetailView(View):
-    template_name = 'tv/watch_channel.html'
-
-    def get(self, request, *args, **kwargs):
-        channel = None
-        try:
-            channel = TvChannelModel.filter_cache(
-                channel_name_slug = kwargs['channel_name_slug']
-                ).first()
-            
-        except :
-            va_logger.error(traceback.format_exc())
-            # TODO : renders ops page
-        
-        return render(request, self.template_name, context= {
-            'channel' : channel
-        })
